+++ mapie/estimator/interface.py
@@ -1,14 +1,12 @@
 from __future__ import annotations
 
 from abc import ABCMeta, abstractmethod
-from typing import Optional, Tuple, Union
-
-from sklearn.base import RegressorMixin
+from typing import Tuple, Union
 
 from mapie._typing import ArrayLike, NDArray
 
 
-class EnsembleEstimator(RegressorMixin, metaclass=ABCMeta):
+class EnsembleEstimator(metaclass=ABCMeta):
     """
     This class implements methods to handle the training and usage of the
     estimator. This estimator can be unique or composed by cross validated
@@ -20,9 +18,7 @@ class EnsembleEstimator(RegressorMixin, metaclass=ABCMeta):
         self,
         X: ArrayLike,
         y: ArrayLike,
-        sample_weight: Optional[ArrayLike] = None,
-        groups: Optional[ArrayLike] = None,
-        **fit_params
+        **kwargs
     ) -> EnsembleEstimator:
         """
         Fit the base estimator under the ``single_estimator_`` attribute.
@@ -30,70 +26,15 @@ class EnsembleEstimator(RegressorMixin, metaclass=ABCMeta):
         and rearrange them into a list, the ``estimators_`` attribute.
         Out-of-fold conformity scores are stored under
         the ``conformity_scores_`` attribute.
-
-        Parameters
-        ----------
-        X: ArrayLike of shape (n_samples, n_features)
-            Input data.
-
-        y: ArrayLike of shape (n_samples,)
-            Input labels.
-
-        sample_weight: Optional[ArrayLike] of shape (n_samples,)
-            Sample weights. If None, then samples are equally weighted.
-            By default ``None``.
-
-        groups: Optional[ArrayLike] of shape (n_samples,)
-            Group labels for the samples used while splitting the dataset into
-            train/test set.
-            By default ``None``.
-
-        **fit_params : dict
-            Additional fit parameters.
-
-        Returns
-        -------
-        EnsembleRegressor
-            The estimator fitted.
         """
 
     @abstractmethod
     def predict(
         self,
         X: ArrayLike,
-        ensemble: bool = False,
-        return_multi_pred: bool = True
+        **kwargs
     ) -> Union[NDArray, Tuple[NDArray, NDArray, NDArray]]:
         """
         Predict target from X. It also computes the prediction per train sample
         for each test sample according to ``self.method``.
-
-        Parameters
-        ----------
-        X: ArrayLike of shape (n_samples, n_features)
-            Test data.
-
-        ensemble: bool
-            Boolean determining whether the predictions are ensembled or not.
-            If ``False``, predictions are those of the model trained on the
-            whole training set.
-            If ``True``, predictions from perturbed models are aggregated by
-            the aggregation function specified in the ``agg_function``
-            attribute.
-
-            If ``cv`` is ``"prefit"`` or ``"split"``, ``ensemble`` is ignored.
-
-            By default ``False``.
-
-        return_multi_pred: bool
-            If ``True`` the method returns the predictions and the multiple
-            predictions (3 arrays). If ``False`` the method return the
-            simple predictions only.
-
-        Returns
-        -------
-        Tuple[NDArray, NDArray, NDArray]
-            - Predictions
-            - The multiple predictions for the lower bound of the intervals.
-            - The multiple predictions for the upper bound of the intervals.
         """